Move node draining util functions under utils.go
